Refactored request.py add added permissions checking

+++ source/plugins/request.py
@@ -2,32 +2,46 @@
 
 import json
 import re
+import os.path
+from discord.utils import get
 
-# TODO: Create function for interacting with the request_data file by taking a lambda to perform on the data
-# TODO: Look for other ways to refactor this code to be more understandable
 
 REQUEST_COMMAND = "request"
-SHOW_COMMAND = "showrequests"
+REQUESTS_COMMAND = "requests"
 RESPOND_COMMAND = "respond"
 
 REQUEST_DATA_FILE_PATH = "data/request_data.json"
 DEFAULT_RESPONSE = "Thank you for submitting your request, the admins will review it shortly"
+NO_PERMISSIONS_RESPONSE = "You do not have the proper permissions to perform this action"
+ADMIN_CHANNEL_NAME = "admins"
+ELEVATED_PERMISSIONS_ROLES = ['Admin', 'Moderator']
 
 
 async def command_request(client, message):
     try:
         request = create_request_item(message)
         add_request(request)
+        await alert_admins(client, message, request)
         await client.send_message(message.channel, DEFAULT_RESPONSE)
     except Exception as error:
         await client.send_message(message.channel, str(error))
         raise error
 
 
-async def command_show_requests(client, message):
+async def command_requests(client, message):
     try:
-        resp = get_requests_list_message()
-        await client.send_message(message.channel, resp)
+        if not user_has_elevated_permissions(message.author):
+            await client.send_message(message.channel, NO_PERMISSIONS_RESPONSE)
+            return
+        
+        command = parse_sub_command(message.content)
+
+        if  command['subcommand'] == 'show':
+            await client.send_message(message.channel, get_requests_list_message())
+        elif command['subcommand'] == 'clear':
+            clear_requests()
+            await client.send_message(message.channel, 'The requests queue has been cleared')
+
     except Exception as error:
         await client.send_message(message.channel, str(error))
         raise error
@@ -35,12 +49,24 @@ async def command_show_requests(client, message):
 
 async def command_respond(client, message):
     try:
+        if not user_has_elevated_permissions(message.author):
+            await client.send_message(message.channel, NO_PERMISSIONS_RESPONSE)
+            return
+
         cmd = parse_respond_command(message.content)
         remove_request(cmd['request_index'])
         await send_respond_response(client, cmd)
+        await client.send_message(message.channel, 'The response has been sent')
     except Exception as error:
         await client.send_message(message.channel, str(error))
         raise error
+
+
+async def alert_admins(client, message, request):
+    admin_channel = get(message.server.channels, name=ADMIN_CHANNEL_NAME)
+    resp = 'New request "{}" by **{}**'.format(request['message'], request['author_name'])
+    await client.send_message(admin_channel, resp)
+
 
 
 async def send_respond_response(client, cmd):
@@ -67,20 +93,52 @@ def parse_respond_command(content):
     }
 
 
+def parse_sub_command(content):
+    parts = content.split(' ')
+    if len(parts) <= 1:
+        raise Exception(
+            'Requests sub-command not given. Run `!requests help` for usage details')
+    
+    return {
+        'subcommand': parts[1],
+    }
+
+
+def user_has_elevated_permissions(user):
+    user_elevated_permissions = [role.name for role in user.roles if role.name in ELEVATED_PERMISSIONS_ROLES]
+
+    return user_elevated_permissions != []
+    
+
 def add_request(request):
-    with open(REQUEST_DATA_FILE_PATH, 'r+') as file:
-        requests = json.load(file)
-        file.seek(0)
-        file.write(json.dumps(requests + [request]))
-        file.truncate()
+    map_over_requests(lambda reqs: [request] + reqs)
 
 
 def remove_request(index):
+    map_over_requests(lambda reqs: remove_element(reqs, index))
+
+
+def clear_requests():
+    map_over_requests(lambda _ : [])
+
+
+def remove_element(l, idx):
+    return l[:idx] + l[idx+1:]
+
+
+def map_over_requests(f):
+    if not requests_file_exists():
+        with open(REQUEST_DATA_FILE_PATH, 'w') as file:
+            requests = f([])
+            file.write(json.dumps(requests))
+
+        return
+
     with open(REQUEST_DATA_FILE_PATH, 'r+') as file:
         requests = json.load(file)
-        del requests[index]
         file.seek(0)
-        file.write(json.dumps(requests))
+        updated_requests = f(requests)
+        file.write(json.dumps(updated_requests))
         file.truncate()
 
 
@@ -112,5 +170,12 @@ def request_display_string(index, request):
 
 
 def get_requests():
+    if not requests_file_exists():
+        return []
+
     with open(REQUEST_DATA_FILE_PATH, 'r') as requests_data_file:
         return json.load(requests_data_file)
+
+
+def requests_file_exists():
+    return os.path.isfile(REQUEST_DATA_FILE_PATH)
